Создает модули и добавляет функцию для генерации комментариев
