+++ js/data.js
@@ -1,0 +1,29 @@
+const NAMES =
+['Артём',
+  'Мария',
+  'Иван',
+  'Ольга',
+  'Дмитрий',
+  'Елена',
+  'Сергей',
+  'Анна',
+  'Павел',
+  'Ирина'
+];
+
+const MESSAGES = [
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Захватывающий закат на пляже, где небо окрашено в потрясающие оттенки оранжевого и розового.',
+  'Горная тропа, ведущая к вершине, где открывается панорамный вид на бескрайние леса и озера.',
+  'Весёлые друзья, смеющиеся и наслаждающиеся пикником в живописном парке.',
+  'Милый котенок, играющий с клубком пряжи, создающий очаровательное зрелище.',
+  'Ночной город с яркими огнями, отражающимися в реке, создавая волшебную атмосферу.'
+];
+
+const SIMILAR_PHOTO_COUNT = 25;
+
+export {NAMES, MESSAGES, DESCRIPTIONS, SIMILAR_PHOTO_COUNT};
